Some Quest Changes + Script Changes
Still working out how i want scripts + quests to work together. I may
change this in the future.

+++ Darkages.Server/Network/ServerFormats/ServerFormat48.cs
@@ -12,7 +12,7 @@
 
         public override void Serialize(NetworkPacketWriter writer)
         {
-            writer.Write((byte)byte.MinValue);
+            writer.Write(byte.MinValue);
         }
     }
 }
